Enforce max number of groups on the server
