+++ src/lib/client/state.svelte.ts
@@ -2,15 +2,6 @@ type Box<T> = {
     value: T;
 };
 
-export type GroupDetails = {
-    groupId: number;
-    groupName: string;
-    members: {
-        netid: string;
-        displayname: string;
-    }[];
-};
-
 type Course = {
     id: string;
     code: string;
