+++ src/routes/api/groups/join/+server.ts
@@ -1,9 +1,10 @@
+import { MAX_GROUPS } from "$lib/constants";
 import { httpCodes } from "$lib/httpCodes";
 import { CASClient } from "$lib/server/cas";
 import { db } from "$lib/server/db";
 import { groupMembers, groups } from "$lib/server/db/schema";
 import type { RequestHandler } from "@sveltejs/kit";
-import { eq } from "drizzle-orm";
+import { count, eq } from "drizzle-orm";
 
 export const POST: RequestHandler = async ({ locals, request }) => {
     CASClient.check(locals.session.data);
@@ -16,7 +17,23 @@ export const POST: RequestHandler = async ({ locals, request }) => {
         });
     }
 
-    await db.transaction(async (tx) => {
+    const txRes = await db.transaction(async (tx) => {
+        const userGroups = await tx
+            .select({
+                count: count()
+            })
+            .from(groupMembers)
+            .where(eq(groupMembers.userId, locals.session.data.netid));
+
+        if (userGroups[0].count >= MAX_GROUPS) {
+            return new Response(
+                "You are already in the maximum number of groups.",
+                {
+                    status: httpCodes.error.badRequest
+                }
+            );
+        }
+
         // Check if group exists
         const group = await tx
             .select()
@@ -35,6 +52,8 @@ export const POST: RequestHandler = async ({ locals, request }) => {
             groupId
         });
     });
+
+    if (txRes instanceof Response) return txRes;
 
     return new Response("User joined group.", {
         status: httpCodes.success.ok
